refactor: retructure entity #49
