+++ src/main/java/com/kospot/domain/multiGame/game/entity/MultiRoadViewGame.java
@@ -1,9 +1,9 @@
 package com.kospot.domain.multiGame.game.entity;
 
-import com.kospot.domain.coordinate.entity.Coordinate;
 import com.kospot.domain.game.entity.GameMode;
 import com.kospot.domain.game.entity.GameType;
 import com.kospot.domain.multiGame.gameRoom.entity.GameRoom;
+import com.kospot.domain.multiGame.gameRound.entity.RoadViewGameRound;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -30,12 +30,12 @@ public class MultiRoadViewGame extends MultiGame {
     private GameRoom gameRoom;
     
     @OneToMany(mappedBy = "multiRoadViewGame", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<GameRound> gameRounds = new ArrayList<>();
+    private List<RoadViewGameRound> roadViewGameRounds = new ArrayList<>();
     
     // 생성 메서드
     public static MultiRoadViewGame createGame(GameRoom gameRoom, GameType gameType, PlayerMatchType matchType, 
                                              Integer roundCount, Boolean increasingDifficulty) {
-        MultiRoadViewGame game = MultiRoadViewGame.builder()
+        return MultiRoadViewGame.builder()
                 .gameType(gameType)
                 .matchType(matchType)
                 .gameMode(GameMode.ROADVIEW)  // 로드뷰 모드로 고정
@@ -45,13 +45,11 @@ public class MultiRoadViewGame extends MultiGame {
                 .increasingDifficulty(increasingDifficulty)
                 .gameRoom(gameRoom)
                 .build();
-                
-        return game;
     }
     
     // 게임 라운드 추가
-    public void addGameRound(GameRound gameRound) {
-        this.gameRounds.add(gameRound);
-        gameRound.setMultiRoadViewGame(this);
+    public void addGameRound(RoadViewGameRound roadViewGameRound) {
+        this.roadViewGameRounds.add(roadViewGameRound);
+        roadViewGameRound.setMultiRoadViewGame(this);
     }
 }
