+++ src/main/java/com/kospot/domain/multiGame/gameResult/entity/GameResult.java
@@ -1,7 +1,9 @@
-package com.kospot.domain.multiGame.game.entity;
+package com.kospot.domain.multiGame.gameResult.entity;
 
 import com.kospot.domain.auditing.entity.BaseTimeEntity;
 import com.kospot.domain.game.entity.GameMode;
+import com.kospot.domain.multiGame.game.entity.MultiPhotoGame;
+import com.kospot.domain.multiGame.game.entity.MultiRoadViewGame;
 import com.kospot.domain.multiGame.gamePlayer.entity.GamePlayer;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
