+++ src/main/java/com/kospot/domain/multiGame/gameRound/entity/RoadViewGameRound.java
@@ -1,7 +1,10 @@
-package com.kospot.domain.multiGame.game.entity;
+package com.kospot.domain.multiGame.gameRound.entity;
 
 import com.kospot.domain.auditing.entity.BaseTimeEntity;
 import com.kospot.domain.coordinate.entity.Coordinate;
+import com.kospot.domain.multiGame.game.entity.MultiRoadViewGame;
+import com.kospot.domain.multiGame.submittion.entity.RoadViewPlayerSubmission;
+import com.kospot.domain.multiGame.submittion.entity.TeamSubmission;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -17,7 +20,7 @@ import java.util.List;
 @SuperBuilder
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class GameRound extends BaseTimeEntity {
+public class RoadViewGameRound extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -29,13 +32,13 @@ public class GameRound extends BaseTimeEntity {
     @JoinColumn(name = "multi_road_view_game_id")
     private MultiRoadViewGame multiRoadViewGame;
     
-    // 라운드에 사용되는 좌표 (정답)
+    // 라운드에 사용되는 정답 좌표
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "coordinate_id")
     private Coordinate targetCoordinate;
     
     @OneToMany(mappedBy = "gameRound", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<PlayerSubmission> playerSubmissions = new ArrayList<>();
+    private List<RoadViewPlayerSubmission> roadViewPlayerSubmissions = new ArrayList<>();
     
     @OneToMany(mappedBy = "gameRound", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<TeamSubmission> teamSubmissions = new ArrayList<>();
@@ -47,14 +50,14 @@ public class GameRound extends BaseTimeEntity {
         this.multiRoadViewGame = multiRoadViewGame;
     }
     
-    public void addPlayerSubmission(PlayerSubmission submission) {
-        this.playerSubmissions.add(submission);
-        submission.setGameRound(this);
+    public void addPlayerSubmission(RoadViewPlayerSubmission submission) {
+        this.roadViewPlayerSubmissions.add(submission);
+        submission.setRoadViewGameRound(this);
     }
     
     public void addTeamSubmission(TeamSubmission submission) {
         this.teamSubmissions.add(submission);
-        submission.setGameRound(this);
+        submission.setRoadViewGameRound(this);
     }
     
     public void finishRound() {
@@ -62,8 +65,8 @@ public class GameRound extends BaseTimeEntity {
     }
     
     // 생성 메서드
-    public static GameRound createRound(Integer roundNumber, Coordinate targetCoordinate) {
-        return GameRound.builder()
+    public static RoadViewGameRound createRound(Integer roundNumber, Coordinate targetCoordinate) {
+        return RoadViewGameRound.builder()
                 .roundNumber(roundNumber)
                 .targetCoordinate(targetCoordinate)
                 .isFinished(false)
